+++ index.php
@@ -6,11 +6,11 @@ $container = new \Rammewerk\Component\Container\Container( true );
 
 
 
-$class = $container->create( \Rammewerk\Component\Container\Tests\TestData\TestClassF::class, ['the variable is here'] );
+$class = $container->create( \Rammewerk\Component\Container\Tests\TestData\TestParamA::class, ['the variable is here'] );
 
 echo 'Let us inspect the container<br>';
 
-echo $class->get();
+#echo $class->get();
 
 #echo $class->get();