+++ src/Error/ContainerException.php
@@ -2,8 +2,7 @@
 
 namespace Rammewerk\Component\Container\Error;
 
+use Psr\Container\ContainerExceptionInterface;
 use RuntimeException;
 
-class ContainerException extends RuntimeException {
-
-}
+class ContainerException extends RuntimeException implements ContainerExceptionInterface {}