+++ tests/TestData/TestParamA.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rammewerk\Component\Container\Tests\TestData;
+
+class TestParamA {
+
+    public function __construct(
+        TestClassB&TestClassC $test
+    ) {}
+
+
+
+}